Show tiles when no nests are present
When focusing on a single nest if that a bird is not detected at the
nest site on the selected date the RGB tile wasn't shown because the
data frame that the tile ids where harvested from is empty.

This change gets the tile ids before filtering the nests to avoid this
issue.