Show nest location on all dates
For understanding the behavior of bird-bird-bird it is useful to be able to see
the identified nest locations across the entire time-series, not only when there
is a bird identified on the date being visualized.

This change selects a single location for each bird, the first location for each
target_ind, and plots that location on all dates.