Add consecutive detections to nest detector
In cases with fewer repeated flights or fast nest failures it can be
useful to count consecutive bird detections as indicative of a nest
with lower counts than for all detections over the time-series.
This adds the ability to do this and applies it to Everglades bird
detections with a default of a single pair of consecutive detections,
two back-to-back bird detections, identified as a nest.

Also updates the nest detections data using this new algorithm.