Add nest detector from imagery sample generation
Randomly pulls up to 25 detected nests and 25 random birds from each
site/year combination. This allows for unbiased checking of if detected
nests are accurate.

Cases where known_nest = "no" should be checked because many
birds are on nests leading to a decent chance of "no" being a real nest.