Only select birds that are not associate with nests for comparison
To compare to nests that had been detected we had been using a set of
locations randomly selected from all bird locations. This still produced
some "non-nest" locations that were part of nests because a lot of birds
are associated with nests. Building on new functionality in the nest
detector this PR filters to only choose from birds that are not
associated with any nest.