Enforce correct types when combining nest files
We write empty nest files for sites with no nest (typically due to
insufficient sampling). These files are given a schema with types, but
these don't appear to be preserved properly resulting in correct types
for the combined object. This change sets the appropriate type when
loading each shape file.